+++ src/features/Settings/SettingsPage/RulesTable/CustomCells/style.module.scss
